pyproject.toml
[tool.poetry]
name = "pymortgage-calc"
version = "0.1.0"
description = ""
authors = ["cdeever <cdeever@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
